+++ src/lib/editor/Editor.cpp
@@ -188,11 +188,10 @@ void Editor::applyFreebasicTheme() {
     SetILexer(FBSciLexer::Create());
 
     // Apply keywords
-    const auto& groups = m_ctx.getConfigManager().keywords().at("groups");
-    for (std::size_t grp = 0; grp < kThemeKeywordGroupsCount; grp++) {
-        wxString key;
-        key.Printf("group%zu", grp + 1);
-        SetKeyWords(static_cast<int>(grp), groups.get_or(key, "").Lower());
+    const auto& keywords = m_ctx.getConfigManager().keywords();
+    for (std::size_t idx = 0; idx < kThemeKeywordCategories.size(); idx++) {
+        const auto key = getThemeCategoryName(kThemeKeywordCategories[idx]);
+        SetKeyWords(static_cast<int>(idx), keywords.get_or(wxString(key), "").Lower());
     }
 
     for (const auto cat : kThemeCategories) {